pyproject.toml
[tool.poetry]
name = "intensity"
version = "0.1.0"
description = "A general discord bot with many fun features"
authors = ["Abhishek Kashyap <abhik5567@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
nextcord = "^2.0.0a3"
numpy = "1.21.4"
pandas = "1.3.4"
python-dotenv = "^0.19.1"
pymongo = "3.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
